refactor: Move hook to new file
